Replace Switch with if/else
The toggleDisplay function stores the id of the button that was clicked into a variable. Then it loops through the charInfo array and checks to see if the button id and current charInfo div id both equal one character (peter, chris, or brian). If so, it gets a reference to the charInfo div's id and toggles the class 'isHidden' for that div.