added fading effect
I added rules to the class 'isHidden' that transitions the height and opacity to zero rather than 'display; none'. I added float: left and rules to transition the opacity to 1 and the height to 230px on the class 'charInfo'. This gives a fading effect when you click the character's name.